x-pack/platform/plugins/shared/automatic_import_v2/server/agents/evaluation/gepa/requirements.txt
gepa>=0.1.0
requests>=2.28.0
python-dotenv>=1.0.0
litellm>=1.0.0
boto3>=1.28.57
